% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{available_fonts}
\alias{available_fonts}
\title{Fonts available in the sticker package.}
\usage{
available_fonts(verbose = TRUE)
}
\arguments{
\item{verbose}{A \code{logical} controlling verbosity.}
}
\value{
A \code{character} with full names to fonts available in
    the package.
}
\description{
Provides a list of available fonts.
}
\examples{
available_fonts()
fnts <- available_fonts(verbose = FALSE)
fnts
}
\author{
Laurent Gatto
}

