% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sticker.R
\name{make_sticker}
\alias{make_sticker}
\title{Create a BioC sticker.}
\usage{
make_sticker(x, package = "Bioconductor", text_size = 6,
  col_text = "#FFFFFF", col_background = "#1881C2",
  col_border = "#87B13F", grob_xmin = 0.3, grob_xmax = 1.65,
  grob_ymin = 0.25, grob_ymax = 1.3, text_x = 1, text_y = 1.4,
  filename = paste(package, "png", sep = "."))
}
\arguments{
\item{x}{The package logo. Can either be a ggplot, lattice or grob
object.}

\item{package}{Package name, to be used as label on the
sticker. Default is "MyPackage".}

\item{text_size}{Label font size. Default is 22.}

\item{col_text}{Label font colour. Default is \code{"#FFFFFF"}.}

\item{col_background}{Background colour. Default is
\code{"#1881C2"}.}

\item{col_border}{Border colour. Default is \code{"#87B13F"}.}

\item{grob_xmin}{Left position of the image. Default is -0.03.}

\item{grob_xmax}{Right position of the image. Default is 2.}

\item{grob_ymin}{Bottom position of the image. Default is 0.05.}

\item{grob_ymax}{Top position of the image. Default is 1.75.}

\item{text_x}{Text horizontal alignment. Default is 1.}

\item{text_y}{Text vertical alignment. Default is 1.44.}

\item{filename}{By default, the sticker is saved saved printed to
a png file named after the \code{package} name. If you prefer
not to create any file, set to \code{NULL}.}
}
\value{
Invisibly returns the \code{ggplot} sticker
    object. Creates a png file as side effect, unless
    \code{filename} is \code{NULL}.
}
\description{
This function can be used to create a hexagonal sticker following
the guidelines described in the BioC-sticker sticker
repository. The function requires the following packages: ggplot2,
ggforce, ggtree, showtext and grid, if lattice objects are used.
}
\examples{
stkr <- make_sticker(package = "Bioconductor",
                     filename = NULL)
print(stkr)
library("ggplot2")
p <- ggplot(aes(x = mpg, y = wt), data = mtcars) + geom_point()
x <- make_sticker(p, "Bioconductor", filename = NULL,
                  grob_xmin = 0.5, grob_xmax = 1.5,
                  grob_ymin = .35, grob_ymax = 1.25)
x
}
\author{
Laurent Gatto and Guangchuang Yu
}

